refactor: move tests for toHaveALengthOf assertion into its own test file
